# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  api.ApiResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  api.muteConvReq:
    properties:
      convId:
        type: string
    required:
    - convId
    type: object
...
  api.revokeMsgReq:
    properties:
      convId:
        type: string
      msgId:
        type: integer
    required:
    - convId
    - msgId
    type: object
...
  api.updateSettingsReq:
    properties:
      msg_preview_enabled:
        type: boolean
      mute_list:
        type: string
      notification_enabled:
        type: boolean
    type: object
  api.uploadAvatarResponse:
    properties:
      file_path:
        type: string
      size:
        type: integer
      url:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: GoIM 即时通讯系统 API — 支持私聊、群聊、朋友圈动态。
  title: GoIM API
  version: "1.0"
paths:
...
  /msg/{msgID}:
    delete:
      description: 删除指定的消息
      parameters:
      - description: 消息ID
        in: path
        name: msgID
        required: true
        type: string
      - description: 会话ID
        in: query
        name: convId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 删除消息
      tags:
      - 消息操作
  /msg/revoke:
    post:
      consumes:
      - application/json
      description: 在2分钟内撤回已发送的消息，仅发送者可撤回
      parameters:
      - description: 撤回请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.revokeMsgReq'
      produces:
      - application/json
      responses:
        "200":
          description: 撤回成功
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "400":
          description: 消息不可撤回
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "403":
          description: 非发送者
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 撤回消息
      tags:
      - 消息操作
  /msg/search:
    get:
      description: 根据关键词搜索用户的消息记录
      parameters:
      - description: 搜索关键词
        in: query
        name: q
        required: true
        type: string
      - description: 返回数量限制
        in: query
        name: limit
        type: integer
      - description: 分页偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 搜索结果
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 搜索消息
      tags:
      - 消息操作
  /settings:
    get:
      description: 获取当前用户的通知偏好和免打扰设置
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/api.ApiResponse'
            - properties:
                data:
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: 获取用户设置
      tags:
      - 设置
    put:
      consumes:
      - application/json
      description: 更新当前用户的通知偏好和免打扰设置
      parameters:
      - description: 设置参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.updateSettingsReq'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 更新用户设置
      tags:
      - 设置
  /settings/mute:
    post:
      consumes:
      - application/json
      description: 将指定会话添加到免打扰列表
      parameters:
      - description: 会话ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.muteConvReq'
      produces:
      - application/json
      responses:
        "200":
          description: 添加成功
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "409":
          description: 会话已存在
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 添加免打扰会话
      tags:
      - 设置
  /settings/mute/{convID}:
    delete:
      description: 将指定会话从免打扰列表中移除
      parameters:
      - description: 会话ID
        in: path
        name: convID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 移除成功
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api.ApiResponse'
        "404":
          description: 会话未找到
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 移除免打扰会话
      tags:
      - 设置
  /upload/avatar:
    post:
      consumes:
      - multipart/form-data
      description: 上传用户头像图片，支持 jpg/png/gif/webp 格式
      parameters:
      - description: 头像图片文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/api.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/api.uploadAvatarResponse'
              type: object
        "400":
          description: 文件格式不支持或文件过大
          schema:
            $ref: '#/definitions/api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 上传头像
      tags:
      - 文件
securityDefinitions:
  BearerAuth:
    description: 在 Authorization 头中携带 JWT 访问令牌，格式：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"